requirements.txt
asgiref==3.7.2
Django
sqlparse==0.4.4
psycopg2-binary==2.9.6
djangorestframework
drf-spectacular==0.26.2
flake8==6.0.0
pre-commit==3.3.3
djangorestframework-simplejwt==5.2.2
pycoingecko==3.1.0